chore: tidy compose_predicate
