+++ src/composer/mod.rs
@@ -8,3 +8,4 @@ pub mod case_converter;
 pub mod knowledge_base;
 pub mod conclusions;
 pub mod function_def;
+mod compose_predicate;
